requirements.txt
requests==2.31
discordwebhook==1.0.3
python-dotenv==1.0.1